+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## vNext
+## v2.1.0 released 2022-07-07
 * Add GitAnnexAPI method getBuildFlags
 * Support Git 2.36.0 features
     * Add BranchOptions --recurse-submodules